requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
playwright==1.41.0
pymupdf==1.23.8
python-multipart==0.0.6
httpx==0.26.0
google-generativeai==0.3.2
python-dotenv==1.0.0
snowflake-connector-python==3.6.0
